Separated out the Custom ValueResolver unit-test.
