+++ tests/Our.Umbraco.Ditto.Tests/Our.Umbraco.Ditto.Tests.csproj
@@ -213,6 +213,7 @@
   <ItemGroup>
     <Compile Include="BasicMappingTests.cs" />
     <Compile Include="ConversionHandlerTests.cs" />
+    <Compile Include="CustomValueResolverTests.cs" />
     <Compile Include="GlobalValueResolverTests.cs" />
     <Compile Include="ValueResolverContextTests.cs" />
     <Compile Include="CurrentContentTests.cs" />
